feat: use internal minio instead of internal nfs-server for workdir
This changes up.sh to use shrinkwraps rather than helm install, as helm install cannot ensure proper ordering between creation/deletion of controllers and resources.

TODO: for now, we have an explicit workdir Dataset for each namespace. We need to change this so that the controller adds it if missing.

Use busybox cp for clone.sh due to oddities with GNU coreutils cp (error setting/changing permissions)

add torchx subshell timeout as in some cases it may hang

+++ platform/prerender.sh
@@ -11,9 +11,12 @@ set -o pipefail
 SCRIPTDIR=$(cd $(dirname "$0") && pwd)
 cd "$SCRIPTDIR"
 
-SCHEDULER_PLUGINS=v0.27.8
-rm -rf scheduler-plugins
-git clone https://github.com/kubernetes-sigs/scheduler-plugins.git --no-checkout --filter=blob:none -b $SCHEDULER_PLUGINS scheduler-plugins
-(cd scheduler-plugins && \
-     git sparse-checkout set --cone manifests && \
-     git checkout $SCHEDULER_PLUGINS)
+if [[ ! -e scheduler-plugins ]]
+then
+    SCHEDULER_PLUGINS=v0.27.8
+    rm -rf scheduler-plugins
+    git clone https://github.com/kubernetes-sigs/scheduler-plugins.git --no-checkout --filter=blob:none -b $SCHEDULER_PLUGINS scheduler-plugins
+    (cd scheduler-plugins && \
+         git sparse-checkout set --cone manifests && \
+         git checkout $SCHEDULER_PLUGINS)
+fi
